+++ chef/environments/development.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 name "development"
 description "development"
 cookbook "draupnir", "= 0.1.0"