+++ models/access_token.go
@@ -1,5 +1,0 @@
-package models
-
-type AccessToken struct {
-	Token string `jsonapi:"attr,token"`
-}