+++ routes/images_test.go
@@ -42,7 +42,6 @@ func TestGetImage(t *testing.T) {
 	}
 
 	assert.Equal(t, http.StatusOK, recorder.Code)
-	assert.Equal(t, []string{mediaType}, recorder.HeaderMap["Content-Type"])
 	assert.Equal(t, append(expected, byte('\n')), recorder.Body.Bytes())
 }
 
@@ -63,7 +62,6 @@ func TestGetImageWhenAuthenticationFails(t *testing.T) {
 	handler.ServeHTTP(recorder, req)
 
 	assert.Equal(t, http.StatusUnauthorized, recorder.Code)
-	assert.Equal(t, []string{mediaType}, recorder.HeaderMap["Content-Type"])
 }
 
 func TestListImages(t *testing.T) {
@@ -96,7 +94,6 @@ func TestListImages(t *testing.T) {
 	}
 
 	assert.Equal(t, http.StatusOK, recorder.Code)
-	assert.Equal(t, []string{mediaType}, recorder.HeaderMap["Content-Type"])
 	assert.Equal(t, append(expected, byte('\n')), recorder.Body.Bytes())
 }
 
@@ -136,7 +133,6 @@ func TestCreateImage(t *testing.T) {
 	}
 
 	assert.Equal(t, http.StatusCreated, recorder.Code)
-	assert.Equal(t, []string{mediaType}, recorder.HeaderMap["Content-Type"])
 	assert.Equal(t, append(expected, byte('\n')), recorder.Body.Bytes())
 }