+++ spec/image_spec.rb
@@ -1,22 +1,23 @@
 # frozen_string_literal: true
+
 require "spec_helper"
 
-RSpec.describe '/images' do
+RSpec.describe "/images" do
   let(:post_payload) do
     {
       data: {
-        type: 'images',
+        type: "images",
         attributes: {
           backed_up_at: timestamp.iso8601,
           anonymisation_script: "CREATE DATABASE foo;",
-        }
-      }
+        },
+      },
     }
   end
 
   let(:timestamp) { Time.utc(2016, 1, 2, 3, 4, 5) }
 
-  describe 'POST /images' do
+  describe "POST /images" do
     let(:timestamp) { Time.utc(2016, 1, 2, 3, 4, 5) }
 
     context "with an invalid shared secret" do
@@ -45,7 +46,7 @@ RSpec.describe '/images' do
     context "with a valid shared secret" do
       let(:secret) { "thesharedsecret" }
 
-      it 'creates an image and serialises it as a response' do
+      it "creates an image and serialises it as a response" do
         timestamp = Time.utc(2016, 1, 2, 3, 4, 5)
         response = post("/images", post_payload)
 
@@ -59,18 +60,18 @@ RSpec.describe '/images' do
               "backed_up_at" => timestamp.iso8601,
               "ready" => false,
               "created_at" => String,
-              "updated_at" => String
-            )
-          }
+              "updated_at" => String,
+            ),
+          },
         )
       end
     end
   end
 
-  describe 'GET /images' do
+  describe "GET /images" do
     before { post("/images", post_payload) }
 
-    it 'returns a JSON payload listing all the images' do
+    it "returns a JSON payload listing all the images" do
       response = get("/images")
 
       expect(response.code).to eq(200)
@@ -84,20 +85,20 @@ RSpec.describe '/images' do
               "backed_up_at" => timestamp.iso8601,
               "ready" => false,
               "created_at" => String,
-              "updated_at" => String
-            )
-          }
-        ]
+              "updated_at" => String,
+            ),
+          },
+        ],
       )
     end
   end
 
-  describe 'GET /images/:id' do
+  describe "GET /images/:id" do
     let!(:image_id) do
-      JSON.parse(post("/images", post_payload))['data']['id']
+      JSON.parse(post("/images", post_payload))["data"]["id"]
     end
 
-    it 'returns a JSON payload showing the image' do
+    it "returns a JSON payload showing the image" do
       response = get("/images/#{image_id}")
 
       expect(response.code).to eq(200)
@@ -110,19 +111,19 @@ RSpec.describe '/images' do
             "backed_up_at" => timestamp.iso8601,
             "ready" => false,
             "created_at" => String,
-            "updated_at" => String
-          )
-        }
+            "updated_at" => String,
+          ),
+        },
       )
     end
   end
 
-  describe 'DELETE /images/:id' do
+  describe "DELETE /images/:id" do
     let!(:image_id) do
-      JSON.parse(post("/images", post_payload))['data']['id']
+      JSON.parse(post("/images", post_payload))["data"]["id"]
     end
 
-    it 'deletes the image and returns a 204' do
+    it "deletes the image and returns a 204" do
       response = delete("/images/#{image_id}")
 
       expect(response.code).to eq(204)
@@ -131,4 +132,3 @@ RSpec.describe '/images' do
     end
   end
 end
-