+++ spec/spec_helper.rb
@@ -1,16 +1,18 @@
 # frozen_string_literal: true
-require 'rest_client'
-require 'json'
-require 'rspec'
+
+require "rest_client"
+require "json"
+require "rspec"
 
 JSON_CONTENT_TYPE = "application/json"
 SERVER_IP = "192.168.2.3"
 SERVER_ADDR = "https://#{SERVER_IP}"
 DATA_PATH = "/draupnir"
 ACCESS_TOKEN = "the-integration-access-token"
+DRAUPNIR_VERSION = `cat DRAUPNIR_VERSION`.freeze
 
 RSpec.configure do |config|
-  def post(path, payload, headers={})
+  def post(path, payload, headers = {})
     RestClient::Request.execute(
       verify_ssl: false,
       method: :post,
@@ -19,7 +21,8 @@ RSpec.configure do |config|
       headers: {
         content_type: JSON_CONTENT_TYPE,
         authorization: "Bearer #{ACCESS_TOKEN}",
-      }.merge(headers)
+        draupnir_version: DRAUPNIR_VERSION,
+      }.merge(headers),
     )
   end
 
@@ -30,8 +33,9 @@ RSpec.configure do |config|
       url: "#{SERVER_ADDR}#{path}",
       headers: {
         content_type: JSON_CONTENT_TYPE,
-        authorization: "Bearer #{ACCESS_TOKEN}"
-      }
+        authorization: "Bearer #{ACCESS_TOKEN}",
+        draupnir_version: DRAUPNIR_VERSION,
+      },
     )
   end
 
@@ -43,12 +47,13 @@ RSpec.configure do |config|
       headers: {
         content_type: JSON_CONTENT_TYPE,
         authorization: "Bearer #{ACCESS_TOKEN}",
-      }
+        draupnir_version: DRAUPNIR_VERSION,
+      },
     )
   end
 
   def destroy_all_instances!
-    instances = JSON.parse(get("/instances"))['data']
+    instances = JSON.parse(get("/instances"))["data"]
 
     instances.each do |instance|
       delete("/instances/#{instance['id']}")
@@ -56,7 +61,7 @@ RSpec.configure do |config|
   end
 
   def destroy_all_images!
-    images = JSON.parse(get("/images"))['data']
+    images = JSON.parse(get("/images"))["data"]
 
     images.each do |image|
       delete("/images/#{image['id']}")
@@ -72,5 +77,4 @@ RSpec.configure do |config|
     destroy_all_instances!
     destroy_all_images!
   end
-
 end